Add mermaidjs asset files and custom css to the theme
Earlier, mermaidjs assets(js and css) files were expected
to be added in the site static resources, which creates an
additional step in the site level. We can add these files
at the theme level so site does not need any additonal
configuration.

Adding the custom css file to display the home page titles
in two columns as part of the theme.